Split tests for response headers
These tests all inspect headers in the response from the CDN. They are
mostly for information and debugging purposes.

This commit is just a copy/paste with the imports adjusted appropriately.
I'll rename the tests separately.
